fixed type casts

+++ ft_strdup.c
@@ -3,9 +3,9 @@
 char	*ft_strdup(const char *str)
 {
 	char	*dup;
-	int	len;
+	size_t	len;
 
-	len = ft_strlen(str) + 1;
+	len = ft_strlen(str) + (size_t)1;
 	dup = (char *)malloc(sizeof(char) * len);
 	if (dup)
 		ft_strcpy(dup, str);
